마리아 DB로 jdbc 변경

+++ src/main/java/kr/co/chunjae/cart/controller/CartController.java
@@ -11,10 +11,10 @@ import org.springframework.web.servlet.ModelAndView;
 
 public interface CartController {
 	public ModelAndView myCartMain(HttpServletRequest request, HttpServletResponse response)  throws Exception;
-	public @ResponseBody String addGoodsInCart(@RequestParam("goods_id") Long goodsId, HttpServletRequest request, HttpServletResponse response)  throws Exception;
-	public  @ResponseBody String modifyCartQty(@RequestParam("goods_id") Long goodsId, @RequestParam("cart_goods_qty") int cartGoodsQty,
+	public @ResponseBody String addGoodsInCart(@RequestParam("goods_id") int goodsId, HttpServletRequest request, HttpServletResponse response)  throws Exception;
+	public  @ResponseBody String modifyCartQty(@RequestParam("goods_id") int goodsId, @RequestParam("cart_goods_qty") int cartGoodsQty,
                                                HttpServletRequest request, HttpServletResponse response)  throws Exception;
-	public ModelAndView removeCartGoods(@RequestParam("cart_id") Long cartId, HttpServletRequest request, HttpServletResponse response)  throws Exception;
+	public ModelAndView removeCartGoods(@RequestParam("cart_id") int cartId, HttpServletRequest request, HttpServletResponse response)  throws Exception;
